- Alignment display: Each nucleotid rectangle is now pre-painted and copied into the actual display canvas   (instead of repainted again and again)   Sequence and paritition dimming as well as mouse over highlighting is now done by transparent overlaying   (instead of repainting the panel or buffer on each action)   Also added a light color gradient to the nuc. rectangles - Fixed update checker (version.txt in /www, upload to webserver with next release)   (already done in previous svn commit) - Fixed a bug in "remove bootstrap value" dialog - Added ToolTips to various table headers - Removed some unneccessary System.out.println()s - Added RF tree distances to model selection webservice - Quick Tree Estimation: Now allows setting of ts/tv and alpha parameter - Fixed bug in model test: AIC2 and BIC score was wrong - Removed 'Optimize...' checkboxes from phyml dialog (not needed at the moment) - Fixed bug: AA sub. models have 0 degress of freedom not 19 - Updated treedist to newer version (needed for model test)

+++ src/topali/var/tree/NHTreeUtils.java
@@ -48,7 +48,7 @@ public class NHTreeUtils
 	 * @param threshold
 	 * @return
 	 */
-	public static String removeBootstrapValues(String tree, int threshold) {
+	public static String removeBootstrapValues(String tree, double threshold) {
 		int s = 0;
 		int e = 0;
 		while(true) {
@@ -60,7 +60,7 @@ public class NHTreeUtils
 			if(e<=(s+1))
 				continue;
 			
-			int bs = Integer.parseInt(tree.substring(s+1, e));
+			double bs = Double.parseDouble(tree.substring(s+1, e));
 			if(bs<threshold) {
 				tree = tree.substring(0, s+1)+tree.substring(e);
 			}
@@ -76,8 +76,8 @@ public class NHTreeUtils
 	 * @param tree
 	 * @return
 	 */
-	public static int[] analyzeBootstrapValues(String tree) {
-		int min = Integer.MAX_VALUE, max = Integer.MIN_VALUE, avg = 0;
+	public static double[] analyzeBootstrapValues(String tree) {
+		double min = Integer.MAX_VALUE, max = Integer.MIN_VALUE, avg = 0;
 		
 		int s = 0;
 		int e = 0;
@@ -91,7 +91,7 @@ public class NHTreeUtils
 			if(e<=(s+1))
 				continue;
 			
-			int bs = Integer.parseInt(tree.substring(s+1, e));
+			double bs = Double.parseDouble(tree.substring(s+1, e));
 			if(bs<min) {
 				min = bs;
 			}
@@ -102,6 +102,6 @@ public class NHTreeUtils
 			c++;
 		}
 		
-		return new int[] {min,(int)((double)avg/(double)c),max};
+		return new double[] {min, avg/(double)c,max};
 	}
 }
